requirements.txt
fastapi==0.95.0
uvicorn==0.21.1
pybind11==2.10.4
pyzmq==25.0.2
msgpack==1.0.5
numpy==1.24.2
pydantic==1.10.7
websockets==11.0.1
requests==2.28.2
pytest==7.2.2
jinja2==3.1.2
python-multipart==0.0.6
